fix: Resolve bug with rich text editor not loading
The issue occurred due to the enhanced navigation feature being used in the button to create a new blog post.

The URL generated by enhanced navigation triggered a fetch request in the network tab of the browser's developer tools, which intercepted the request.

This caused the JavaScript scripts required for the rich text editor to not load properly, while the CSS loaded successfully.

To resolve this, a complete round trip to the server for server-side rendering was implemented.